Refactor updateQuestionResponses
